+++ lib/src/classifier/logistic_regressor/gradient_logistic_regressor.dart
@@ -1,12 +1,11 @@
-import 'package:ml_algo/src/classifier/classifier.dart';
 import 'package:ml_algo/src/classifier/linear_classifier_mixin/linear_classifier_mixin.dart';
+import 'package:ml_algo/src/classifier/logistic_regressor/logistic_regressor.dart';
 import 'package:ml_algo/src/cost_function/cost_function_type.dart';
 import 'package:ml_algo/src/helpers/add_intercept.dart';
 import 'package:ml_algo/src/optimizer/gradient/learning_rate_generator/learning_rate_type.dart';
 import 'package:ml_algo/src/optimizer/initial_weights_generator/initial_weights_type.dart';
 import 'package:ml_algo/src/optimizer/optimizer_factory.dart';
 import 'package:ml_algo/src/optimizer/optimizer_factory_impl.dart';
-import 'package:ml_algo/src/optimizer/optimizer_type.dart';
 import 'package:ml_algo/src/score_to_prob_mapper/score_to_prob_mapper.dart';
 import 'package:ml_algo/src/score_to_prob_mapper/score_to_prob_mapper_factory.dart';
 import 'package:ml_algo/src/score_to_prob_mapper/score_to_prob_mapper_factory_impl.dart';
@@ -16,8 +15,9 @@ import 'package:ml_linalg/dtype.dart';
 import 'package:ml_linalg/matrix.dart';
 import 'package:ml_linalg/vector.dart';
 
-class LogisticRegressor with LinearClassifierMixin implements Classifier {
-  LogisticRegressor(
+class GradientLogisticRegressor with LinearClassifierMixin
+    implements LogisticRegressor {
+  GradientLogisticRegressor(
       this.trainingFeatures,
       this.trainingOutcomes, {
         // public arguments
@@ -29,7 +29,6 @@ class LogisticRegressor with LinearClassifierMixin implements Classifier {
         int batchSize = 1,
         bool fitIntercept = false,
         double interceptScale = 1.0,
-        OptimizerType optimizer = OptimizerType.gradient,
         LearningRateType learningRateType = LearningRateType.constant,
         InitialWeightsType initialWeightsType = InitialWeightsType.zeroes,
         Matrix initialWeights,
@@ -48,12 +47,11 @@ class LogisticRegressor with LinearClassifierMixin implements Classifier {
         scoreToProbMapper =
           scoreToProbMapperFactory.fromType(_scoreToProbMapperType, dtype),
         classLabels = trainingOutcomes.uniqueRows(),
-        weightsByClasses = optimizerFactory.fromType(
-          optimizer,
+        weightsByClasses = optimizerFactory.gradient(
           addInterceptIf(fitIntercept, trainingFeatures, interceptScale),
           trainingOutcomes,
           dtype: dtype,
-          costFunctionType: CostFunctionType.logLikelihood,
+          costFnType: CostFunctionType.logLikelihood,
           scoreToProbMapperType: _scoreToProbMapperType,
           learningRateType: learningRateType,
           initialWeightsType: initialWeightsType,